requirements.txt
google-genai
langchain-community
langchain
PyPDF2
chromadb
sentence-transformers